canonical selector hashes
